requirements.txt
aiohttp
python-dotenv
google-generativeai
textstat
streamlit
